+++ lib/rubocop/cop/style/negated_if.rb
@@ -8,6 +8,8 @@ module Rubocop
       class NegatedIf < Cop
         include NegativeConditional
 
+        MSG = 'Favor `%s` over `%s` for negative conditions.'
+
         def on_if(node)
           return unless node.loc.respond_to?(:keyword)
           return if node.loc.keyword.is?('elsif')
@@ -15,8 +17,12 @@ module Rubocop
           check(node)
         end
 
-        def error_message
-          'Favor `unless` over `if` for negative conditions.'
+        def message(node)
+          if node.loc.keyword.is?('if')
+            format(MSG, 'unless', 'if')
+          else
+            format(MSG, 'if', 'unless')
+          end
         end
 
         private
@@ -26,7 +32,10 @@ module Rubocop
             condition, _body, _rest = *node
             # unwrap the negated portion of the condition (a send node)
             pos_condition, _method, = *condition
-            corrector.replace(node.loc.keyword, 'unless')
+            corrector.replace(
+              node.loc.keyword,
+              node.loc.keyword.is?('if') ? 'unless' : 'if'
+            )
             corrector.replace(condition.loc.expression,
                               pos_condition.loc.expression.source)
           end